Renamed old Cog class to Degree

+++ src/main/java/com/wicio/shiplog/log/api/dto/CreateLogRequest.java
@@ -8,7 +8,7 @@ public record CreateLogRequest(
     Double XCoordinate,
     Double speedOverGround,
     Integer courseOverGround,
-    Double windDirection,
+    Integer windDirection,
     Double windSpeed,
     Boolean stationary) {
